+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Connector/Tasklet/PimEnterprisePrepareProductsCriteriaEvaluationTasklet.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Connector\Tasklet;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Application\ProductEvaluation\CreateMissingCriteriaEvaluationsInterface;
-use Akeneo\Pim\Automation\DataQualityInsights\Domain\Repository\CriterionEvaluationRepositoryInterface;
 use Akeneo\Tool\Component\Connector\Step\TaskletInterface;
 use Psr\Log\LoggerInterface;
 
@@ -30,22 +29,16 @@ final class PimEnterprisePrepareProductsCriteriaEvaluationTasklet implements Tas
     /** @var LoggerInterface */
     private $logger;
 
-    /** @var CriterionEvaluationRepositoryInterface */
-    private $productCriterionEvaluationRepository;
-
     public function __construct(
         CreateMissingCriteriaEvaluationsInterface $createMissingProductsCriteriaEvaluations,
-        LoggerInterface $logger,
-        CriterionEvaluationRepositoryInterface $productCriterionEvaluationRepository
+        LoggerInterface $logger
     ) {
         $this->createMissingProductsCriteriaEvaluations = $createMissingProductsCriteriaEvaluations;
         $this->logger = $logger;
-        $this->productCriterionEvaluationRepository = $productCriterionEvaluationRepository;
     }
 
     public function execute(): void
     {
-        $this->cleanCriteriaOfDeletedProducts();
         $this->createMissingCriteriaEvaluations();
     }
 
@@ -64,10 +57,5 @@ final class PimEnterprisePrepareProductsCriteriaEvaluationTasklet implements Tas
                 ]
             );
         }
-    }
-
-    private function cleanCriteriaOfDeletedProducts()
-    {
-        $this->productCriterionEvaluationRepository->deleteUnknownProductsEvaluations();
     }
 }